+++ crates/ruborist/src/fs.rs
@@ -17,6 +17,8 @@ use std::future::Future;
 use std::io;
 use std::path::{Path, PathBuf};
 
+use anyhow::Context;
+
 /// Read a file and return its contents as a UTF-8 string.
 pub async fn read_to_string(path: impl AsRef<Path>) -> io::Result<String> {
     tokio_fs_ext::read_to_string(path).await
@@ -30,10 +32,13 @@ pub async fn exists(path: impl AsRef<Path>) -> bool {
 /// Read a JSON file and deserialize it.
 ///
 /// Uses `read` + `simd_json::from_slice` for SIMD-accelerated parsing.
-pub async fn read_json<T: serde::de::DeserializeOwned>(path: impl AsRef<Path>) -> io::Result<T> {
-    let mut bytes = tokio_fs_ext::read(path).await?;
+pub async fn read_json<T: serde::de::DeserializeOwned>(path: impl AsRef<Path>) -> anyhow::Result<T> {
+    let path = path.as_ref();
+    let mut bytes = tokio_fs_ext::read(path)
+        .await
+        .context(format!("Failed to read {}", path.display()))?;
     simd_json::serde::from_slice(&mut bytes)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        .context(format!("Failed to parse {}", path.display()))
 }
 
 /// Glob pattern matching trait.
