+++ crates/ruborist/src/lib.rs
@@ -22,6 +22,7 @@
 //! }).await?;
 //! ```
 
+pub mod fs;
 pub mod model;
 pub mod resolver;
 pub mod service;
